requirements.txt
tensorflow==2.19.0
fastapi
uvicorn
pillow
opencv-python
numpy